requirements.txt

numpy==1.18.1
regex==2020.7.14
tensorflow==2.1.0
tensorflow-hub==0.7.0
tweepy==3.8.0
vaderSentiment==3.3.2
transformers==3.0.2
pandas
textblob 
tweet-preprocessor
nltk
matplotlib
seaborn 
plotly
psutil
kaleido
geograpy3
locatiotagger

